Fix something small
