+++ .gitignore
@@ -1,4 +1,5 @@
 # Projects
+0-*
 .vscode/
 datasets/*Speech*
 logdir/
