--- melgan/identity.py
@@ -1,12 +1,0 @@
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-
-
-class Identity(nn.Module):
-    def __init__(self):
-        super(Identity, self).__init__()
-
-    def forward(self, x):
-        return x
-
