+++ vocoder/res_stack.py
